% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{abi}
\alias{abi}
\title{Calculate P obs for a FLStock object.}
\usage{
abi(object, age, ...)
}
\arguments{
\item{object}{A FLStock object.}

\item{age}{Reference ages obtained from abiAge.}
}
\value{
An FLQuant object containing P obs.
}
\description{
This function calculates P obs for a FLStock object.
}
