% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{asa}
\alias{asa}
\title{amat: age at a specific maturity}
\usage{
asa(object, ...)
}
\arguments{
\item{object}{An FLQuant object with an ogive with the proportion of older spawners,
by default calculated by soawnOnce}
}
\value{
Returns the proportion of old spawners by biomass.
}
\description{
based on interpolation between values or a specuific value
}
\examples{
\dontrun{
data(ple4)
amat(ple4)}
}
\seealso{
\code{\link{ssb}}, \code{\link{ssb.age}},
\code{\link{pos}}, \code{\link{spawnOnce}},
\code{\link{asa}}, \code{\link{awa}},
\code{\link{ssb}}, \code{\link{amat}}, \code{\link{wmat}},
\code{\link{fjuv}}, \code{\link{fapex}}
}
