% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTrend.R
\name{analyzeLinearTrend}
\alias{analyzeLinearTrend}
\title{Analyze Linear Trend}
\usage{
analyzeLinearTrend(x, y, varName)
}
\arguments{
\item{x}{Predictor variable (usually year)}

\item{y}{Response variable (SSB or F)}

\item{varName}{Name of variable for reporting}
}
\value{
List with trend analysis results
}
\description{
Analyze Linear Trend
}
