% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTrend.R
\name{analyzeMannKendallTrend}
\alias{analyzeMannKendallTrend}
\title{Analyze Mann-Kendall Trend}
\usage{
analyzeMannKendallTrend(y, varName)
}
\arguments{
\item{y}{Response variable (SSB or F)}

\item{varName}{Name of variable for reporting}
}
\value{
List with trend analysis results
}
\description{
Analyze Mann-Kendall Trend
}
