% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex.R
\name{aopt}
\alias{aopt}
\title{aopt()}
\usage{
aopt(object, nyears = 3)
}
\arguments{
\item{object}{class FLStock}
}
\value{
FLQuant with annual spr0y
}
\description{
Function to compute Aopt, the age where an unfished cohort attains maximum biomass
}
\author{
Henning Winker (JRC)
}
