% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{awa}
\alias{awa}
\alias{fjuv}
\title{Age based indicators}
\usage{
awa(object, ...)

fjuv(object, ...)
}
\description{
the average Weight at age anomaly are the deviations in the expected weight
of fish at a certain age, which can be influenced by environmental conditions,
food availability, and fishing pressure. Changes in average weight at age can
affect the productivity and sustainability of fish stocks. This concept is
important for understanding growth patterns and population dynamics
To calculate ASW anomalies, you first calculate the average weight at
age for your stock and then identify deviations from this average over time.
}
\examples{
\dontrun{
data(ple4)
fjuv(ple4)
}
}
\seealso{
\code{\link{ssb}}, \code{\link{ssb.age}},
\code{\link{pos}}, \code{\link{spawnOnce}},
\code{\link{asa}}, \code{\link{awa}},
\code{\link{ssb}}, \code{\link{amat}}, \code{\link{wmat}},
\code{\link{fjuv}}, \code{\link{fapex}}
}
