% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{benchmark}
\alias{benchmark}
\alias{benchmark,FLStock-method}
\alias{benchmark,FLStocks-method}
\alias{benchmark,FLBRP-method}
\title{benchmark}
\usage{
benchmark(object, ...)

\S4method{benchmark}{FLStock}(object)

\S4method{benchmark}{FLStocks}(object)

\S4method{benchmark}{FLBRP}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments (not currently used)}
}
\value{
An FLPar object containing benchmark reference points (Fmsy, Flim, Fpa, Blim, Bpa, Btrigger)
}
\description{
A generic function to extract benchmark reference points from FLStock objects
}
