% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{bheqz}
\alias{bheqz}
\title{bheqz, Beverton & Holt
z = (k * (linf - lmean)) / (lmean - lc)
lmean = sum(naa * len) / sum(naa)
lc, length at first capture}
\usage{
bheqz(x, linf, k, t0, lc = lc50(x))
}
\description{
bheqz, Beverton & Holt
z = (k * (linf - lmean)) / (lmean - lc)
lmean = sum(naa * len) / sum(naa)
lc, length at first capture
}
\examples{
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
}
