% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBRP-blim.R
\name{blim,FLBRP-method}
\alias{blim,FLBRP-method}
\title{Calculate Biomass Limit Reference Points}
\usage{
\S4method{blim}{FLBRP}(object, ratio = 0.3)
}
\arguments{
\item{object}{An FLBRP object containing stock-recruitment parameters}

\item{ratio}{Numeric value between 0 and 1 indicating the fraction of virgin recruitment
to use for Blim calculation (default: 0.3)}
}
\value{
An FLPar object containing Blim reference points with the following quantities:
\itemize{
\item harvest: Fishing mortality at Blim
\item yield: Catch at Blim
\item rec: Recruitment at Blim
\item ssb: Spawning Stock Biomass at Blim
\item biomass: Total biomass at Blim
\item revenue: Revenue at Blim
\item cost: Cost at Blim
\item profit: Profit at Blim
}
}
\description{
Calculates biomass limit reference points (Blim) based on a ratio of virgin recruitment
for an FLBRP object
}
\details{
The method:
\itemize{
\item Extracts virgin recruitment from reference points
\item Applies the specified ratio to calculate Blim recruitment
\item Computes associated reference points (harvest, yield, SSB, biomass, etc.)
}
}
\examples{
\dontrun{
data(ple4)
brp <- FLBRP(ple4)
blim_ref <- blim(brp, ratio=0.3)
}

}
