% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{calcPriors,FLBRP-method}
\alias{calcPriors,FLBRP-method}
\title{Calculate priors for an FLBRP object}
\usage{
\S4method{calcPriors}{FLBRP}(object)
}
\arguments{
\item{object}{An object of class FLBRP}
}
\value{
A named vector of calculated prior values
}
\description{
This method calculates priors for an FLBRP object based on reference points and observations.
}
\examples{
\dontrun{
  eql=FLBRP(...)  # Create or load an FLBRP object
  calcPriors(eql)
}
}
