% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptRebuild.R
\name{calculateBmsyKRatioAlternative}
\alias{calculateBmsyKRatioAlternative}
\title{Calculate BMSY/K ratio for alternative Pella-Tomlinson parameterization}
\usage{
calculateBmsyKRatioAlternative(p)
}
\arguments{
\item{p}{Shape parameter (can be positive or negative)}
}
\value{
BMSY/K ratio
}
\description{
Calculates BMSY/K ratio for the alternative P-T model that allows negative p values
}
