% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptRebuild.R
\name{calculateInitialBiomassAlternative}
\alias{calculateInitialBiomassAlternative}
\title{Calculate initial biomass for alternative Pella-Tomlinson model}
\usage{
calculateInitialBiomassAlternative(nyrs, rEff, p, B1, K)
}
\arguments{
\item{nyrs}{Number of years for recovery}

\item{rEff}{Effective growth rate}

\item{p}{Shape parameter (negative)}

\item{B1}{Target biomass}

\item{K}{Carrying capacity}
}
\value{
Initial biomass required
}
\description{
Calculates initial biomass required for recovery in given time
for the alternative P-T parameterization with negative p values.
}
