% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptRebuild.R
\name{calculateRecoveryTime}
\alias{calculateRecoveryTime}
\title{Calculate recovery time using Pella-Tomlinson dynamics}
\usage{
calculateRecoveryTime(B0, B1, K, rEff, p, invalid, naMask)
}
\arguments{
\item{B0}{Initial biomass}

\item{B1}{Target biomass}

\item{K}{Carrying capacity}

\item{rEff}{Effective growth rate}

\item{p}{Shape parameter}

\item{invalid}{Logical vector of invalid cases}

\item{naMask}{Logical vector of NA cases}
}
\value{
Recovery time vector
}
\description{
Calculate recovery time using Pella-Tomlinson dynamics
}
