% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTrend.R
\name{classifyStockTrend}
\alias{classifyStockTrend}
\title{Classify Stock Trend}
\usage{
classifyStockTrend(ssbTrend, fTrend, pThreshold)
}
\arguments{
\item{ssbTrend}{SSB trend analysis results}

\item{fTrend}{F trend analysis results}

\item{pThreshold}{Significance threshold}
}
\value{
List with trend classification
}
\description{
Classify Stock Trend
}
