% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{createDataFrameFromTimeSeries}
\alias{createDataFrameFromTimeSeries}
\title{Create data frame from time series data}
\usage{
createDataFrameFromTimeSeries(biomass_ts, catch_ts)
}
\description{
Create data frame from time series data
}
\keyword{internal}
