% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{createResultFLPar}
\alias{createResultFLPar}
\title{Create result FLPar object}
\usage{
createResultFLPar(opt_result, biomass_data, yield_data)
}
\description{
Create result FLPar object
}
\keyword{internal}
