% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-eqPlot.R
\name{curveSS,list-method}
\alias{curveSS,list-method}
\title{Extracts yield and surplus production from Stock Synthesis output
processes Stock Synthesis output to calculate yield and surplus production.}
\usage{
\S4method{curveSS}{list}(object)
}
\arguments{
\item{object}{An output from SS_outputs}
}
\value{
A list containing processed time series data, equilibrium yield,
reference points, and triangle data.
}
\description{
Extracts yield and surplus production from Stock Synthesis output
processes Stock Synthesis output to calculate yield and surplus production.
}
\examples{
# Assuming 'ss_output' is an instance of StockSynthesisOut#' # result <- curveSS(ss_output)
}
