% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-funcs.R
\name{eb.obs}
\alias{eb.obs}
\alias{eb.obs,FLBRP-method}
\title{Get Observed Exploitable Biomass from FLBRP Object}
\usage{
eb.obs(x, ...)

\S4method{eb.obs}{FLBRP}(x)
}
\arguments{
\item{x}{An object of class \code{FLBRP}}

\item{...}{Additional arguments (not used)}
}
\value{
FLQuant containing observed exploitable biomass or NULL if the
attribute doesn't exist
}
\description{
Retrieves observed exploitable biomass values stored as an
attribute in an FLBRP object.
}
\examples{
\dontrun{
data(ple4brp)
eb.obs(ple4brp)
}

}
\seealso{
\code{\link{ssb.obs}} for observed spawning stock biomass
}
