% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesdata-utils.R
\name{eqsimFn}
\alias{eqsimFn}
\title{Internal Function for EqSim Reference Point Extraction}
\usage{
eqsimFn(object)
}
\arguments{
\item{object}{An FLStock object}
}
\value{
An FLPar object
}
\description{
Internal Function for EqSim Reference Point Extraction
}
