% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{eqsimsFn}
\alias{eqsimsFn}
\title{Internal Function for EqSim Reference Point Extraction}
\usage{
eqsimsFn(x)
}
\arguments{
\item{x}{An FLStock object}
}
\value{
An FLPar object
}
\description{
Internal Function for EqSim Reference Point Extraction
}
