% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex.R
\name{fabs}
\alias{fabs}
\title{Compute instantaneous F, such that  F_a = Fabs*Sa}
\usage{
fabs(stk, nyears = 5, year = NULL)
}
\arguments{
\item{stk}{Input FLStock object.}

\item{nys}{numbers of last years to compute selectivity}
}
\value{
value Fabs
}
\description{
Compute instantaneous F, such that  F_a = Fabs*Sa
}
\author{
Henning Winker (JRC)
}
