% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{fishlifesFn}
\alias{fishlifesFn}
\title{Internal Function for FishLife Parameter Extraction}
\usage{
fishlifesFn(x)
}
\arguments{
\item{x}{An FLStock object}
}
\value{
An FLPar object
}
\description{
Internal Function for FishLife Parameter Extraction
}
