% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{wmat}
\alias{wmat}
\title{Fjuv}
\usage{
wmat(object, ...)
}
\description{
The ratio of juvenile to apical fishing mortality and compares
the fishing mortality rates of juvenile fish to those at the peak of the fishing mortality
curve (fapex). It's used to assess the impact of fishing on different life
stages of fish and to inform management strategies that protect juvenile
fish to ensure the sustainability of fish stocks.
The ratio of juvenile to apical fishing mortality (Fjuv/Fapical)
This ratio involves calculating fishing mortality rates for juvenile and
apical age classes separately and then dividing them.
}
\examples{
\dontrun{
data(ple4)
fjuv(ple4)}
}
\seealso{
\code{\link{ssb}}, \code{\link{ssb.age}},
\code{\link{pos}}, \code{\link{spawnOnce}},
\code{\link{asa}}, \code{\link{awa}},
\code{\link{ssb}}, \code{\link{amat}}, \code{\link{wmat}},
\code{\link{fjuv}}, \code{\link{fapex}}
}
