% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCandy-package.R
\docType{package}
\name{FLCandy-package}
\alias{FLCandy-package}
\alias{FLCandy}
\title{FLCandy Package}
\description{
Candidate Methods for FLR - Contains prototypes and experimental methods for the FLR framework
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flr/FLCandy}
  \item Report bugs at \url{https://github.com/flr/FLCandy/issues}
}

}
\author{
\strong{Maintainer}: Laurence Kell \email{laurie@kell.es}

}
\keyword{internal}
