% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{FLifeParFn}
\alias{FLifeParFn}
\title{Internal Function for FLife Parameter Extraction}
\usage{
FLifeParFn(x)
}
\arguments{
\item{x}{An FLStock object}
}
\value{
An FLPar object with life history parameters
}
\description{
Internal Function for FLife Parameter Extraction
}
