% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{getInitialParameters}
\alias{getInitialParameters}
\title{Get initial parameter estimates}
\usage{
getInitialParameters(biomass_data, yield_data)
}
\description{
Get initial parameter estimates
}
\keyword{internal}
