% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length2.R
\name{Growth models}
\alias{Growth models}
\alias{vonbert}
\alias{ivonbert}
\alias{gompertz}
\alias{richards}
\title{Growth models}
\usage{
vonbert(linf, k, t0, age)

ivonbert(linf, k, t0, len)

gompertz(linf, a, k, age)

richards(linf, k, b, m, age)
}
\arguments{
\item{linf}{Asymptotic length.}

\item{k}{Growth coefficient.}

\item{t0}{Theoretical age at length zero.}

\item{age}{Vector of ages for which to calculate lengths.}

\item{len}{Vector of lengths for which to calculate ages.}

\item{a}{Parameter a.}

\item{b}{Parameter b.}

\item{m}{Parameter m.}
}
\value{
A vector of lengths corresponding to the given ages.

A vector of ages corresponding to the given lengths.

A vector of lengths corresponding to the given ages.

A vector of lengths corresponding to the given ages.
}
\description{
Functions to model fish growth.

Von Bertalanffy growth model.

Inverse Von Bertalanffy growth model.

Gompertz growth model.

Richards growth model.
}
\details{
This group of functions provides different growth models commonly used in fisheries science.
}
\examples{
data(ple4)
vonbert(linf = 35, k = 0.352, t0 = -0.26, age = 1:14)
vonbert

vonbert(linf = 35, k = 0.352, t0 = -0.26, age = 1:14)
ivonbert(35, 0.352, -0.26, 1:34)
gompertz(linf = 179.13, k = 0.4088, a = 1.7268, age = 1:12)
richards(linf = 178.63, k = 0.424, b = -7.185, m = 2880.4, age = 1:12)
}
