% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{hasNoTimeSeriesData}
\alias{hasNoTimeSeriesData}
\title{Check if time series data has no meaningful dimensions}
\usage{
hasNoTimeSeriesData(biomass_ts, catch_ts)
}
\description{
Check if time series data has no meaningful dimensions
}
\keyword{internal}
