% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haupt2.R
\name{haupt,data.frame,numeric-method}
\alias{haupt,data.frame,numeric-method}
\title{Calculate something using the haupt method for data.frame and numeric objects.}
\usage{
\S4method{haupt}{data.frame,numeric}(
  object,
  pars,
  lc = FLife::vonB(pars["sel1"], pars) * 0.9,
  lmax = "missing",
  ...
)
}
\arguments{
\item{object}{An object of class data.frame.}

\item{pars}{A parameter object (numeric).}

\item{lc}{A threshold value for lc.}

\item{lmax}{A threshold value for lmax.}

\item{...}{Additional arguments (not used in this example).}
}
\value{
The result of the calculation.
}
\description{
This method calculates something using the \code{haupt} function when the input object is of class data.frame and pars is of class numeric.
}
