% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haupt2.R
\name{haupt,FLQuant,FLPar-method}
\alias{haupt,FLQuant,FLPar-method}
\title{Calculate something using the haupt method for FLQuant and FLPar objects.}
\usage{
\S4method{haupt}{FLQuant,FLPar}(
  object,
  pars,
  lc = FLife::vonB(pars["sel1"], pars) * 0.9,
  lmax = "missing",
  ...
)
}
\arguments{
\item{object}{An object of class FLQuant.}

\item{pars}{A parameter object (class FLPar) containing necessary parameters.}

\item{lc}{A threshold value for lc.}

\item{lmax}{A threshold value for lmax.}

\item{...}{Additional arguments (not used in this example).}
}
\value{
The result of the calculation.
}
\description{
This method calculates something using the \code{haupt} function when the input object is of class FLQuant and the parameter object is of class FLPar.
}
