% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-funcs.R
\name{ind}
\alias{ind}
\alias{ind,FLStock,missing-method}
\title{Calculate Fishery Indicators}
\usage{
ind(x, y, ...)

\S4method{ind}{FLStock,missing}(x)
}
\arguments{
\item{x}{An FLStock object containing stock data}

\item{y}{An optional FLBRP object containing reference points (default missing)}
}
\value{
A data.frame of biological indicators including:
\itemize{
\item SSB/SSBmsy
\item F/Fmsy
\item Spawning potential ratio (SPR)
\item Maturity/age structure metrics
}
}
\description{
Computes a suite of biological reference indicators for stock assessment using FLR objects.
}
\examples{
\dontrun{
data(ple4)
brp <- FLBRP(ple4)
ind(ple4) # Without reference points
ind(ple4, brp) # With reference points
}

}
