% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invALK2.R
\name{invALK,FLQuant,missing,missing-method}
\alias{invALK,FLQuant,missing,missing-method}
\title{Inverse Age-Length Key}
\usage{
\S4method{invALK}{FLQuant,missing,missing}(
  object,
  model = vonbert,
  age = ages(object),
  cv = 0.1,
  lmax = 1.2,
  bin = 1,
  max = NULL,
  reflen = NULL
)
}
\arguments{
\item{object}{An FLQuant object containing length-at-age data}

\item{model}{Not used, maintained for method consistency}

\item{age}{Vector of ages (default: ages from object)}

\item{cv}{Coefficient of variation for length distribution (default: 0.1)}

\item{lmax}{Multiplier for maximum length (default: 1.2)}

\item{bin}{Length bin size (default: 1)}

\item{max}{Maximum length (default: NULL)}

\item{reflen}{Reference length for SD calculation (default: NULL)}
}
\value{
An FLPar object containing the probability matrix of age given length
}
\description{
Creates an inverse Age-Length Key from length-at-age data using normal distributions
}
