% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesdata-utils.R
\name{kobe,FLStock,missing-method}
\alias{kobe,FLStock,missing-method}
\title{Extracts Kobe Indicators  (SSB/BMSY and F/FMSY)}
\usage{
\S4method{kobe}{FLStock,missing}(path, method)
}
\arguments{
\item{path}{An object of class FLStock containing stock assessment results}

\item{method}{Not used, included for method consistency}
}
\value{
An FLQuants object containing:
\itemize{
\item stock: Time series of SSB/BMSY ratios
\item harvest: Time series of F/FMSY ratios
}
}
\description{
Calculates SSB and fishing mortality ratios (SSB/BMSY and F/FMSY)
for Kobe plot from an FLStock object.
}
\details{
The method requires that the FLStock object has:
\itemize{
\item eqsim attribute containing BMSY reference point
\item benchmark attribute containing FMSY reference point
}
}
\examples{
\dontrun{
data(ple4)
kobe_indicators <- kobe(ple4)
}

}
\seealso{
\code{\link[FLCore]{FLStock}}, \code{\link[FLCore]{FLQuants}}
}
