% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhz.R
\docType{methods}
\name{bhz}
\alias{bhz}
\alias{lhEql}
\alias{lhEql-method}
\alias{lhEql,FLPar-method}
\title{Derives an \code{FLBRP} from life history parameters}
\usage{
bhz(lfd, par)
}
\arguments{
\item{params}{an \code{FLPar} object with life history parameters}

\item{growth}{function that takes an \code{FLPar} object with parameters, by default \code{vonB}}

\item{m}{\code{character} takes the natural mortality model name, by default gislason}

\item{mat}{function that takes an \code{FLPar} object with parameters, by default \code{logistic}}

\item{sel}{function that takes an \code{FLPar} object with parameters, by default \code{dnormal}}

\item{sr}{\code{character} value, "bevholt" by default}

\item{range}{\code{numeric} with age range by default from 0 to 40}

\item{spwn}{\code{numeric} give propotion of year when spawning occurs, by default is params\link{"a50"}-floor(params\link{"a50"})}

\item{fish}{\code{numeric} give propotion of year when fishing occurs, by default 0.5}

\item{units}{\code{character} for vectors in \code{FLBRP} returned by method}

\item{midyear}{when growth measured, default 0.5}

\item{...}{any other arguments}
}
\value{
\code{FLBRP} object
}
\description{
Takes an \code{FLPar} object with life history and selectivity parameters
and generates a corresponding \code{FLBRP} object. Can use a range of functional forms.
}
\examples{
\dontrun{
load("lfd.1.RData")
load("design.RData")
par=lhs[["Pollachius pollachius"]]
par=rbind(FLPar("lc"=FLife::vonB(par["sel1"],par)),par)
z=bhz(lfd,par)
plot(z)
}



}
\seealso{
\code{\link{lhPar}}, \code{\link{lhRef}}

\code{\link{vonB}} \code{\link{lorenzen}} \code{\link{sigmoid}}
}
