% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{lmaxy}
\alias{lmaxy}
\title{lmaxy, length class with maximum biomass in catch}
\usage{
lmaxy(x, lenwt)
}
\description{
lmaxy, length class with maximum biomass in catch
}
\examples{
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
}
