% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matSel.R
\name{matSel}
\alias{matSel}
\title{Calculate metrics comparing maturity and selectivity patterns}
\usage{
matSel(x)
}
\arguments{
\item{x}{An FLStock object containing catch.sel and mat slots}
}
\value{
A data frame with age-based metrics comparing selectivity and maturity
}
\description{
Calculate metrics comparing maturity and selectivity patterns
}
