% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\docType{methods}
\name{properties}
\alias{properties}
\title{properties}
\usage{
properties(object, ...)
}
\arguments{
\item{object}{an \code{FLBRP} object}
}
\value{
A named vector of calculated prior values

\code{FLPar} with estimates of MSY based reference points
}
\description{
Calculates additional reference points
}
\details{
Calculate priors for an FLBRP object

This method calculates priors for an FLBRP object based on reference points and observations.
}
\examples{
\dontrun{
  eql=FLBRP(...)  # Create or load an FLBRP object
  calcPriors(eql)
}
\dontrun{

data(ple4brp)

fbar(ple4brp)=FLQuant(seq(0,c(refpts(ple4brp)["crash","harvest"]),length.out=101))
refpts(ple4brp)=properties(ple4brp)

plot(ple4brp,ncol=2)
}

}
\seealso{
\code{\link{FLBRP},\link{refpts}}
}
