% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{msy}
\alias{msy}
\alias{msy,FLPar-method}
\title{Calculate MSY from Pella-Tomlinson Parameters}
\usage{
msy(param)

\S4method{msy}{FLPar}(param)
}
\arguments{
\item{param}{FLPar object containing r, bmsy, virgin, and p parameters}
}
\value{
FLPar object containing calculated MSY
}
\description{
Calculates the maximum sustainable yield (MSY) using the
Pella-Tomlinson production model parameters.
}
\examples{
\dontrun{
# Assuming you have an FLPar with required parameters
param=FLPar(r=0.5, bmsy=500, virgin=1000, p=0.5)
msy(param)
}
}
