% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex-indicators.R
\name{osr}
\alias{osr}
\title{Operational Sex Ratio (OSR)}
\usage{
osr(object, ...)
}
\arguments{
\item{object}{An FLStock object with 'F' and 'M' units}

\item{female}{Character specifying female unit name}

\item{male}{Character specifying male unit name}
}
\value{
FLQuant with OSR values (proportion of females)
}
\description{
$ OSR = \frac{\sum_l (N_{1,f} \times Mat_{1,f})}{\sum_l (N_{1,f} \times Mat_{1,f}) + \sum_l (N_{1,m} \times Mat_{1,m})} $
}
