% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{pellaTMleCore}
\alias{pellaTMleCore}
\title{Core MLE function for Pella-Tomlinson parameter estimation}
\usage{
pellaTMleCore(object, biomass = "ssb", ...)
}
\arguments{
\item{object}{List containing biomass and yield vectors}

\item{biomass}{Character string specifying biomass type}

\item{...}{Additional optimization parameters}
}
\value{
FLPar object with estimated parameters
}
\description{
Internal function that performs maximum likelihood estimation
of Pella-Tomlinson model parameters from time series data.
Optimized with vectorized calculations and efficient parameter estimation.
}
\keyword{internal}
