% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{pmega}
\alias{pmega}
\title{pmega, Proportion of individuals above L opt + 10\%}
\usage{
pmega(x, linf, lopt = linf * 2/3)
}
\description{
pmega, Proportion of individuals above L opt + 10\%
}
\examples{
pmega(samps, linf=60)
}
