% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{spawnOnce}
\alias{spawnOnce}
\title{POS: proportion of the SSB made up of older individuals.}
\usage{
spawnOnce(object, ...)
}
\arguments{
\item{object}{An FLQuant object with an ogive with the proportion of older spawners,
by default calculated by soawnOnce}
}
\value{
Returns the proportion of old spawners by biomass.
}
\description{
Older spawners are important for the genetic diversity
and resilience of fish populations, as they tend to produce more, and potentially
higher quality, eggs. To calculate POS, first identify the older age classes,
sum their biomass, then divide by the total spawning stock biomass (SSB).
The older age classes can be identified from the maturity ogive \code{mat}, for example
using the \code{spawnOnce} method. This shifts the maturity by 1 age soi that it idenifies
the proportion of individuals that have spawned before
}
\examples{
\dontrun{
data(ple4)
pos(ple4)}
}
