% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTrend.R
\name{print.stockTrend}
\alias{print.stockTrend}
\title{Print method for stockTrend objects}
\usage{
print.stockTrend(x, ...)
}
\arguments{
\item{x}{stockTrend object}

\item{...}{Additional arguments}
}
\description{
Print method for stockTrend objects
}
