% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors,FLBRP-method}
\alias{priors,FLBRP-method}
\title{Get priors for biomass dynamic assessment from an \code{FLBRP} object}
\usage{
\S4method{priors}{FLBRP}(object)
}
\arguments{
\item{object}{An object of class \code{FLBRP}}
}
\value{
priors attribute as an \code{FLPar} of the \code{FLBRP} object
}
\description{
Get priors for biomass dynamic assessment from an \code{FLBRP} object
}
\examples{
\dontrun{
  eql=FLBRP(...)  # Create or load an FLBRP object
  priors(eql)
}
}
