% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{priors}
\alias{priors}
\title{priors}
\usage{
priors(object)
}
\arguments{
\item{object}{An \code{FLBRP} from which to extract priors}

\item{...}{additional arguments passed to methods}
}
\value{
\code{priors} attribute as an \code{FLPar}
}
\description{
priors for biomass dynamic assessment from an \code{FLBRP} object
}
