% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/FLBRP-processError.R
\name{processError}
\alias{processError}
\alias{processError,FLBRP-method}
\title{Calculate Process Error for FLBRP Object}
\usage{
processError(object, ...)

\S4method{processError}{FLBRP}(object)
}
\arguments{
\item{object}{An object of class FLBRP}
}
\value{
A data frame containing:
\item{ssb}{Observed spawning stock biomass}
\item{catch}{Observed catch}
\item{sp}{Predicted surplus production}
\item{pe}{Process error calculated as (SSB\link{t+1} - SSB\link{t} + Catch\link{t} - SP\link{t})/SSB\link{t}}
}
\description{
Calculates process error from observed SSB and catch data compared to predicted surplus production
}
\note{
Process error quantifies the deviation between observed and predicted population dynamics
}
\examples{
\dontrun{
data(ple4brp)
processError(ple4brp)
}
}
