% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{production}
\alias{production}
\alias{production,FLPar,FLQuant-method}
\alias{production,FLPar,numeric-method}
\alias{production,FLPar,FLPar-method}
\title{Calculate Production Using Pella-Tomlinson Model}
\usage{
production(param, biomass)

\S4method{production}{FLPar,FLQuant}(param, biomass)

\S4method{production}{FLPar,numeric}(param, biomass)

\S4method{production}{FLPar,FLPar}(param, biomass)
}
\arguments{
\item{param}{FLPar object containing r, p, and virgin parameters}

\item{biomass}{FLQuant or numeric vector of biomass values}
}
\value{
Production values (same class as input biomass)
}
\description{
Calculates surplus production (net production) for given biomass
levels using the Pella-Tomlinson surplus production model.
}
\examples{
\dontrun{
# Assuming you have parameters and biomass
param=FLPar(r=0.5, p=0.5, virgin=1000)
biomass=FLQuant(500)
production(param, biomass)
}
}
