% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{production_fast}
\alias{production_fast}
\title{Vectorized production function for speed}
\usage{
production_fast(r, p, k, biomass)
}
\description{
Vectorized production function for speed
}
\keyword{internal}
