% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/pt.R
\name{pt}
\alias{pt}
\alias{pt-method}
\alias{pt,FLPar-method}
\alias{pt,numeric-method}
\alias{pt,FLBRP-method}
\title{Production Type Method for Surplus Production Models}
\usage{
pt(object)

pt(object)

\S4method{pt}{FLBRP}(object)
}
\arguments{
\item{object}{An FLBRP object}

\item{shape}{Ratio of BMSY to virgin biomass (default: object\link{"bmsy"}/object\link{"v"})}

\item{what}{Function to calculate biomass metric (default: ssb)}

\item{...}{Additional arguments passed to methods}
}
\value{
An FLPar object containing:
\itemize{
\item r: intrinsic growth rate
\item k: carrying capacity
\item p: shape parameter
\item msy: maximum sustainable yield
\item bmsy: biomass at MSY
\item fmsy: fishing mortality at MSY
\item virgin: unfished biomass
}

FLPar object containing Pella-Tomlinson parameters
}
\description{
Estimates parameters of surplus production models using the PT method by deriving
parameters from reference points and biomass ratios

Converts an FLBRP object to Pella-Tomlinson surplus production
model parameters (r, p, virgin, bmsy, msy, fmsy, fcrash).
}
\details{
Production Type (PT) Method
}
\examples{
\dontrun{
params <- FLPar(c(bmsy=500, msy=125, v=1000))
pt(params)
}

\dontrun{
# Assuming you have an FLBRP object called 'brp'
pt(brp)
}
}
