% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptRebuild.R
\name{ptParameterizations}
\alias{ptParameterizations}
\title{Get information about Pella-Tomlinson parameterizations}
\usage{
ptParameterizations()
}
\value{
List with information about both parameterizations
}
\description{
Provides information about the two Pella-Tomlinson parameterizations
and helps users understand which one to use.
}
\examples{
ptParameterizations()
}
