% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{r}
\alias{r}
\alias{r,FLPar-method}
\title{Calculate Intrinsic Growth Rate r from Pella-Tomlinson Parameters}
\usage{
r(param)

\S4method{r}{FLPar}(param)
}
\arguments{
\item{param}{FLPar object containing p, msy, bmsy, and virgin parameters}
}
\value{
FLPar object containing calculated r parameter
}
\description{
Calculates the intrinsic growth rate r for the Pella-Tomlinson
surplus production model using the relationship between MSY, BMSY, and p.
}
\examples{
\dontrun{
# Assuming you have an FLPar with required parameters
param=FLPar(p=0.5, msy=100, bmsy=500, virgin=1000)
r(param)
}
}
