% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/FLBRP-rebuild.R
\name{rebuild}
\alias{rebuild}
\alias{rebuild,FLBRP-method}
\title{Get priors for biomass dynamic assessment from an \code{FLBRP} object}
\usage{
rebuild(
  object,
  targetF = NULL,
  targetSSB = NULL,
  nInitial = 100,
  growthRate = 0.25,
  minVal = 0.000001,
  maxVal = 1,
  burnin = 20,
  truncate = TRUE
)

\S4method{rebuild}{FLBRP}(
  object,
  targetF = computeRefpts(object)["msy", "harvest"] * 0,
  targetSSB = computeRefpts(object)["msy", "ssb"],
  nInitial = 100,
  growthRate = 0.25,
  minVal = 0.000001,
  maxVal = 1,
  burnin = 20,
  truncate = TRUE
)
}
\arguments{
\item{object}{An object representing the population}

\item{targetF}{Target fishing mortality during rebuilding}

\item{targetSSB}{Target spawning stock biomass}

\item{nInitial}{Number of initial SSB levels}

\item{growthRate}{Growth rate for depletion sequence}

\item{minVal}{Minimum depletion value}

\item{maxVal}{Maximum depletion value}

\item{burnin}{Number of years for burn-in period}

\item{truncate}{Whether to remove burn-in period}
}
\value{
priors attribute as an \code{FLPar} of the \code{FLBRP} object

An object with rebuilding trajectories
}
\description{
Projects rebuilding trajectories from different initial SSB levels
}
\examples{
\dontrun{
  eql=FLBRP(...)  # Create or load an FLBRP object
  priors(eql)
}
Rebuild a fish population

}
