% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptRebuild.R
\name{rebuildTime}
\alias{rebuildTime}
\alias{rebuildTime,numeric-method}
\alias{rebuildTime,FLPar-method}
\title{Calculate rebuild time using Pella-Tomlinson dynamics}
\usage{
rebuildTime(object, ...)

\S4method{rebuildTime}{numeric}(object, target = NULL, r, p, k, bmsy = NULL, F = 0, ...)

\S4method{rebuildTime}{FLPar}(object, initial, target = NULL, F = 0, ...)
}
\arguments{
\item{object}{Current biomass or FLPar object with parameters}

\item{...}{Additional arguments}

\item{target}{Target biomass (default is BMSY)}

\item{r}{Intrinsic growth rate (if object is numeric)}

\item{p}{Shape parameter (if object is numeric)}

\item{k}{Virgin biomass (if object is numeric)}

\item{bmsy}{Biomass at MSY (if object is numeric)}

\item{F}{Fishing mortality during rebuild (default=0)}
}
\value{
Rebuild time in years
}
\description{
Calculate the time required to rebuild a stock from current biomass to target biomass
using Pella-Tomlinson surplus production dynamics.
}
