% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocFLQ.R, R/roc2.R
\name{roc}
\alias{roc}
\title{Calculate ROC (Receiver Operating Characteristic) statistics for two numeric vectors.}
\usage{
roc(state, indicator, ...)

roc(state, indicator, ...)
}
\arguments{
\item{state}{A numeric vector representing the state values.}

\item{indicator}{A numeric vector representing the indicator values.}
}
\value{
A data frame containing the following columns:
\describe{
\item{state}{The state values.}
\item{label}{A logical vector indicating whether each state is greater than 1 (TRUE) or not (FALSE).}
\item{indicator}{The indicator values.}
\item{TPR}{The True Positive Rate (TPR) calculated as TP / (TP + FN).}
\item{FPR}{The False Positive Rate (FPR) calculated as FP / (FP + TN).}
\item{TP}{The True Positives (TP).}
\item{TN}{The True Negatives (TN).}
\item{FP}{The False Positives (FP).}
\item{FN}{The False Negatives (FN).}
\item{TSS}{The True Skill Score (TSS) calculated as (TP / (TP + FN)) - (FP / (FP + TN)).}
\item{order}{The order of the indicator values after sorting in descending order.}
}
}
\description{
This function calculates ROC statistics, including True Positive Rate (TPR), False Positive Rate (FPR), True Positives (TP), True Negatives (TN), False Positives (FP), False Negatives (FN), and True Skill Score (TSS) for two numeric vectors.
}
\examples{
data(ple4)
# OM 'realities' on stock status relative to refpt
label <- rlnorm(100, log(ssb(ple4))[, '2017'], 0.2) < 850000
# Model estimates of SSB
ind <- rlnorm(100, log(ssb(ple4) * 1.05)[,'2017'], 0.6)
# Compute TSS, returns data.frame
roc(label, ind)
ggplot(roc(label, ind), aes(x=TPR, y=FPR)) +
  geom_line() +
  geom_abline(slope=1, intercept=0, colour="red", linetype=2)
# Multiple years are computed together, but labelled
label <- rlnorm(100, log(ssb(ple4)[, 52:61]), 0.2) < 850000
ind <- rlnorm(100, log(ssb(ple4)[, 52:61] * 1.05), 0.6)
roc(label, ind)
# BUG: BUT years do not work
ggplot(roc(label, ind), aes(x=TPR, y=FPR)) +
  geom_line()
}
