% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{rocFn}
\alias{rocFn}
\title{ROC Curve Coordinates}
\usage{
rocFn(labels, ind)
}
\arguments{
\item{labels}{Binary vector of true labels}

\item{scores}{Numeric vector of prediction scores}
}
\value{
A data frame containing:
\itemize{
\item TPR - True Positive Rate (Sensitivity)
\item FPR - False Positive Rate (1-Specificity)
\item labels - Ordered labels
\item reference - Sorted scores
}
}
\description{
Calculates the coordinates for a Receiver Operating Characteristic (ROC) curve
}
\examples{
\dontrun{
labels <- c(1,0,1,1,0)
scores <- c(0.9, 0.1, 0.8, 0.7, 0.3)
roc_coords <- rocFn(labels, scores)
}

}
