% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex.R
\name{selage}
\alias{selage}
\title{coverts harvest() and catch.sel() selectivity at age Sa}
\usage{
selage(stk, nyears = 5, year = NULL)
}
\arguments{
\item{stk}{Input FLStock object.}

\item{nyears}{numbers of last years to compute selectivity}

\item{year}{option to specify year range, overwrites nyears}
}
\value{
FLQuant of Selectivity-at-age Sa
}
\description{
coverts harvest() and catch.sel() selectivity at age Sa
}
\author{
Henning Winker (JRC)
}
