% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex.R
\name{selex}
\alias{selex}
\title{computes selex curves}
\usage{
selex(Sa, pars, ssq = FALSE)
}
\arguments{
\item{Sa}{selectivity at age Sa = selage(stk)}

\item{pars}{selexpars S50, S95, Smax, Dcv, Dmin}

\item{ssq}{if TRUE return sum-of-squares}
}
\value{
FLquants selex predictions
}
\description{
computes selex curves
}
\author{
Henning Winker (JRC)
}
