% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{shape}
\alias{shape}
\alias{shape,FLPar-method}
\alias{shape,numeric-method}
\title{Calculate BMSY/K Ratio from Shape Parameter p}
\usage{
shape(p)

\S4method{shape}{FLPar}(p)

\S4method{shape}{numeric}(p)
}
\arguments{
\item{p}{FLPar object containing p parameters}
}
\value{
FLPar object containing calculated BMSY/K ratios
}
\description{
Calculates the BMSY/K ratio (shape) for the Pella-Tomlinson surplus
production model given the shape parameter p. This is the inverse of the p() function.
}
\examples{
\dontrun{
# Create FLPar with p parameters
pParams=FLPar(0.5, 1.0, 1.5)
shape(pParams)
}
}
