% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{ssb}
\alias{ssb}
\title{spawnOnce}
\usage{
ssb(object, ...)
}
\arguments{
\item{object}{An FLStock object.}
}
\value{
object An FLQuant object with an ogive with the proportion of older spawners.
}
\description{
Shifts maturity-at-age ogive by 1 age to estimate the propotion
of individuals who have spawned at least once
}
\examples{
data(ple4)
spawnOnce(ple4)
}
\seealso{
\code{\link{ssb}}, \code{\link{ssb.age}},
\code{\link{pos}}, \code{\link{spawnOnce}},
\code{\link{asa}}, \code{\link{awa}},
\code{\link{ssb}}, \code{\link{amat}}, \code{\link{wmat}},
\code{\link{fjuv}}, \code{\link{fapex}}
}
