% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex-indicators.R
\name{srlen}
\alias{srlen}
\title{Example}
\usage{
srlen(object, ...)
}
\arguments{
\item{object}{An FLStock object with 'female' and 'male' units}

\item{female}{Character specifying female unit name}

\item{male}{Character specifying male unit name}
}
\value{
FLQuant with sex ratio values (female proportion) at length
}
\description{
Example
}
\examples{
\dontrun{

data(ple4sex)

## Males are =0, so set to same as females
mat(ple4sex)[,,"M"]=mat(ple4sex)[,,"F"]

index=osr(ple4sex)

ggplot(index) + 
  geom_line(aes(x = year, y = data)) +
  ylab("Operational Sex Ratio (Female Proportion)")
  
 }
Calculate Sex Ratio-at-Length for an FLStock object

$ Sex ratio-at-length = \frac{N_{1,f}}{N_{1,f} + N_{1,m}} \$
    Where:
    
    - \$ N_{1,f} \$: Female numbers at length \$ l \$
    - \$ N_{1,m} \$: Male numbers at length \$ l \$
    
    

}
