% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{ssb,FLStock-method}
\alias{ssb,FLStock-method}
\title{Age based indicators}
\usage{
\S4method{ssb}{FLStock}(object, ...)
}
\description{
Age based indicators
}
\examples{
\dontrun{
library(FLCore)

data(ple4)

awa(ple4)
}

}
