% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{priors}
\alias{priors}
\title{ssb.age}
\usage{
priors(object, ...)
}
\arguments{
\item{object}{An \code{FLBRP} from which to extract priors}

\item{...}{additional arguments passed to methods}
}
\value{
Returns the spawning stock biomass by age as an FLQuant.

\code{priors} attribute as an \code{FLPar}
}
\description{
Calculates the spawning stock biomass by age for an FLStock object.

priors for biomass dynamic assessment from an \code{FLBRP} object
}
\examples{
data(ple4)
ssb.age(ple4)
}
\seealso{
\code{\link{ssb}}
}
