% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{ssb2,FLStock-method}
\alias{ssb2,FLStock-method}
\title{ssb2}
\usage{
\S4method{ssb2}{FLStock}(object, ...)
}
\arguments{
\item{object}{An FLStock object containing stock assessment data.}

\item{...}{additional arguments, to replace slots in the FLStock object.}
}
\value{
Returns the spawning stock biomass as an FLQuant.
}
\description{
Calculates the spawning stock biomass for an FLStock object,
}
\examples{
data(ple4)
FLCore:::ssb2(ple4)
\dontrun{
library(FLCore)

data(ple4)

awa(ple4)
}

}
\seealso{
\code{\link{ssb.age}}
}
