% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/stdz.R
\name{stdz}
\alias{stdz}
\alias{stdz,numeric-method}
\alias{stdz,matrix-method}
\alias{stdz,array-method}
\alias{stdz,FLQuant-method}
\title{Standardize Values}
\usage{
stdz(object, na.rm = TRUE, ...)

\S4method{stdz}{numeric}(object, mn = mean(object, na.rm = TRUE))

\S4method{stdz}{matrix}(object, mn = mean(object, na.rm = TRUE))

\S4method{stdz}{array}(object, mn = mean(object, na.rm = TRUE))

\S4method{stdz}{FLQuant}(object, mn = mean(object, na.rm = TRUE))
}
\arguments{
\item{na.rm}{Logical indicating whether to remove NA values when computing statistics}

\item{x}{A numeric vector to be standardized.}
}
\value{
An object of the same class as the input with standardized values

A numeric vector where each element is standardized (mean = 0, standard deviation = 1).
}
\description{
Standardizes values by subtracting the mean and dividing by the standard deviation.

Standardises a numeric vector by subtracting the mean and dividing by the
standard deviation.
Missing values (\code{NA}) are ignored during computation.
}
\details{
Standardization follows the formula: (x - mean(x))/sd(x)
}
\examples{
\dontrun{
# For numeric vector
x <- 1:10
stdz(object)

# For FLQuant
data(ple4)
standardized_catch <- stdz(catch(ple4))
}

# Standardize a simple numeric vector
vec <- c(1, 2, 3, 4, 5)
stdz(vec)

# Handle missing values
vec_with_na <- c(1, 2, NA, 4, 5)
stdz(vec_with_na)

}
