% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockState.R
\name{stockStateFactor}
\alias{stockStateFactor}
\alias{stockStateFactor,FLStock,missing,missing-method}
\alias{stockStateFactor,missing,FLQuant,FLQuant-method}
\alias{stockStateFactor,data.frame,missing,missing-method}
\title{Create Stock State Factor}
\usage{
stockStateFactor(
  object,
  ssb = NULL,
  f = NULL,
  method = "linear",
  p = 0.05,
  minYears = 5,
  recentYears = NULL,
  levels = c("Decreasing", "Stable", "Increasing")
)

\S4method{stockStateFactor}{FLStock,missing,missing}(
  object,
  ssb = NULL,
  f = NULL,
  method = "linear",
  p = 0.05,
  minYears = 5,
  recentYears = NULL,
  levels = c("Decreasing", "Stable", "Increasing")
)

\S4method{stockStateFactor}{missing,FLQuant,FLQuant}(
  object,
  ssb = NULL,
  f = NULL,
  method = "linear",
  p = 0.05,
  minYears = 5,
  recentYears = NULL,
  levels = c("Decreasing", "Stable", "Increasing")
)

\S4method{stockStateFactor}{data.frame,missing,missing}(
  object,
  ssb = NULL,
  f = NULL,
  method = "linear",
  p = 0.05,
  minYears = 5,
  recentYears = NULL,
  levels = c("Decreasing", "Stable", "Increasing")
)
}
\arguments{
\item{object}{An FLStock object, FLQuant object, data frame, or list of these objects}

\item{ssb}{An FLQuant object for spawning stock biomass (optional if object is FLStock)}

\item{f}{An FLQuant object for fishing mortality (optional if object is FLStock)}

\item{method}{Method for trend analysis}

\item{p}{Significance threshold}

\item{levels}{Order of factor levels (default: c("Decreasing", "Stable", "Increasing"))}

\item{min_years}{Minimum years required}

\item{recent_years}{Number of recent years to focus on}
}
\value{
A factor with stock states
}
\description{
Creates a factor variable for stock states with ordered levels
(Decreasing < Stable < Increasing) for use in statistical analysis.
}
\examples{
\dontrun{
data(ple4)
state_factor=stockStateFactor(ple4)
print(state_factor)  # Factor with levels: Decreasing < Stable < Increasing

# Using FLQuant objects
ssb_quant=ssb(ple4)
f_quant=fbar(ple4)
state_factor=stockStateFactor(ssb=ssb_quant, f=f_quant)
}

}
