% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplusProduction.R
\docType{methods}
\name{surplusProduction}
\alias{surplusProduction}
\title{surplusProduction}
\usage{
surplusProduction(x)
}
\arguments{
\item{object}{an \code{FLStock} object}

\item{seasons}{a numeric with seasons}
}
\value{
\code{FLStock} object
}
\description{
Calculates the surplus production and expected yield etc for the estinates of SSB and biomass
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{expand}}
}
