% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{theme_base}
\alias{theme_base}
\title{Custom Minimal Theme for ggplot2}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
theme_base
}
\arguments{
\item{base_size}{Base font size, default is 12}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top"), default is "none"}
}
\value{
A ggplot2 theme object
}
\description{
Creates a custom theme based on theme_minimal with specific text sizes and no legend by default
}
\examples{
\dontrun{
library(ggplot2)

# Basic usage
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_my()
  
# With legend
ggplot(mtcars, aes(wt, mpg, color =factor(cyl))) +
  geom_point() +
  theme_my(legend_position ="right")
}

}
\keyword{datasets}
