% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/FLBRP-tseries.R,
%   R/FLStock-tseries.R
\docType{methods}
\name{tseries}
\alias{tseries}
\alias{tseries,FLBRP-method}
\alias{tseries,FLStock-method}
\alias{tseries,FLStocks-method}
\title{Extract Time Series Statistics from FLStock Objects}
\usage{
tseries(object, ...)

tseries(object, ...)

\S4method{tseries}{FLBRP}(object)

\S4method{tseries}{FLStock}(
  object,
  flqs = list(catch = function(object) catch(object), eb = function(object)
    ebiomass(object), ssb = function(object) ssb(object), f = function(object)
    fbar(object), h = function(object) catch(object)/ebiomass(object), m =
    function(object)
    FLQuant(aaply(m(object)[ac(range(object)["minfbar"]:range(object)["maxfbar"])], 2,
    mean), dimnames = dimnames(fbar(object))))
)

\S4method{tseries}{FLStocks}(
  object,
  flqs = list(catch = function(object) catch(object), eb = function(object)
    ebiomass(object), ssb = function(object) ssb(object), f = function(object)
    fbar(object), h = function(object) catch(object)/ebiomass(object), m =
    function(object)
    FLQuant(aaply(m(object)[ac(range(object)["minfbar"]:range(object)["maxfbar"])], 2,
    mean), dimnames = dimnames(fbar(object))))
)
}
\arguments{
\item{object}{an \code{FLBRP} object}

\item{...}{Additional arguments (not currently used)}

\item{seasons}{a numeric with seasons}
}
\value{
A data frame containing time series of:
\itemize{
\item catch - Catch values
\item eb - Exploitable biomass
\item ssb - Spawning stock biomass
\item f - Fishing mortality (Fbar)
\item h - Harvest rate (catch/ebiomass)
\item m - Mean natural mortality
}

A data frame containing time series of:
\itemize{
\item catch - Catch values
\item eb - Exploitable biomass
\item ssb - Spawning stock biomass
\item f - Fishing mortality (Fbar)
\item h - Harvest rate (catch/ebiomass)
\item m - Mean natural mortality
}

\code{FLQuants} object
}
\description{
Creates a data frame of time series statistics from FLStock objects, including
catch, ebiomass, SSB, fishing mortality, harvest rate, and mean natural mortality.

Creates a data frame of time series statistics from FLStock objects, including
catch, ebiomass, SSB, fishing mortality, harvest rate, and mean natural mortality.

Calculates the surplus production and expected yield etc for the estinates of SSB and biomass
}
\examples{
\dontrun{
# For single stock
data(ple4)
ts1 <- tseries(ple4)

# For multiple stocks
stocks <- FLStocks(stock1=ple4, stock2=ple4)
ts2 <- tseries(stocks)
}

\dontrun{
# For single stock
data(ple4)
ts1 <- tseries(ple4)

# For multiple stocks
stocks <- FLStocks(stock1=ple4, stock2=ple4)
ts2 <- tseries(stocks)
}

\dontrun{
}
}
\seealso{
\code{\link{expand}}
}
