% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMle.R
\name{validateAndExtractData}
\alias{validateAndExtractData}
\title{Validate and extract data from data frame}
\usage{
validateAndExtractData(object, biomass)
}
\description{
Validate and extract data from data frame
}
\keyword{internal}
