% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abi.R
\name{ABI}
\alias{ABI}
\title{i.e. i) calc ref age, ii) calc P( N ) at FMSY, iii) calc P obs, iv) calc ratio of iii & iii}
\usage{
ABI(x, y, ref = "msy", p = 0.9)
}
\description{
i.e. i) calc ref age, ii) calc P( N ) at FMSY, iii) calc P obs, iv) calc ratio of iii & iii
}
