% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{benchmarksFn}
\alias{benchmarksFn}
\title{Internal Function for Benchmark Extraction}
\usage{
benchmarksFn(x)
}
\arguments{
\item{x}{An FLStock object}
}
\value{
An FLPar object
}
\description{
Internal Function for Benchmark Extraction
}
