% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{eqsim}
\alias{eqsim}
\alias{eqsim,FLStock-method}
\alias{eqsim,FLStocks-method}
\title{eqsim}
\usage{
eqsim(object, ...)

\S4method{eqsim}{FLStock}(object)

\S4method{eqsim}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments (not currently used)}
}
\value{
An FLPar object containing EqSim reference points
}
\description{
A generic function to extract EqSim reference points from FLStock objects
}
