% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{fishlife}
\alias{fishlife}
\alias{fishlife,FLStock-method}
\alias{fishlife,FLStocks-method}
\title{Extract FishLife Parameters}
\usage{
fishlife(object, ...)

fishlife(object, ...)

\S4method{fishlife}{FLStock}(object)

\S4method{fishlife}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments (not currently used)}
}
\value{
An FLPar object containing FishLife parameters

An FLPar object containing FishLife parameters
}
\description{
A generic function to extract FishLife parameters from FLStock objects

A generic function to extract FishLife parameters from FLStock objects
}
