% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCandy-package.R
\docType{package}
\name{FLCandy-package}
\alias{FLCandy}
\alias{FLCandy-package}
\title{FLCandy: Candidate Methods for FLR}
\description{
Contains prototypes and experimental methods for the FLR (Fisheries Library for R) framework. Provides implementations of various fisheries assessment and management tools including biological reference points, harvest control rules, and spatial analysis functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flr/FLCandy}
  \item Report bugs at \url{https://github.com/flr/FLCandy/issues}
}

}
\author{
\strong{Maintainer}: Laurence Kell \email{laurie@kell.es}

}
