% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{FLifePar}
\alias{FLifePar}
\alias{FLifePar,FLStock-method}
\alias{FLifePar,FLStocks-method}
\title{Extract FLife Parameters}
\usage{
FLifePar(object, ...)

\S4method{FLifePar}{FLStock}(object)

\S4method{FLifePar}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments (not currently used)}
}
\value{
An FLPar object containing life history parameters
}
\description{
Generic function to extract FLife parameters from FLStock objects
}
