% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftmb.R
\name{ftmb}
\alias{ftmb}
\title{Fits Stock Recruitment Relationships (SRR) in TBM}
\usage{
ftmb(
  object,
  spr0 = spr0,
  s = 0.7,
  s.est = TRUE,
  s.logitsd = 1.3,
  inits = function(object = object, s = s) c(median(c(log(rec(object)))), log(0.4),
    to_logits(s)),
  lower = function(object = object, s = -20) c(quantile(c(log(rec(object))), probs =
    0.1), log(0.05), -20),
  upper = function(object = object, s = 20) c(quantile(c(log(rec(object))), probs = 0.9),
    log(1.5), 20),
  SDreport = TRUE
)
}
\arguments{
\item{object}{Input FLSR object.}

\item{spr0}{unfished spawning biomass per recruit from FLCore::spr0(FLStock) or FLCore:::spr0Yr}

\item{s}{steepness parameter of SRR (fixed or prior mean)}

\item{s.est}{option to estimate steepness}

\item{s.logitsd}{prior sd for logit(s), default is 1.3 (flat) if s.est = TRUE}

\item{inits}{option to specify initial values of log(r0), log(SigR) and logit(s)}

\item{lower}{option to specify lower bounds of log(r0), log(SigR) and logit(s)}

\item{upper}{option to specify upper bounds of log(r0), log(SigR) and logit(s)}

\item{SDreport}{option to converge hessian and get vcov}
}
\value{
A list containing elements 'FLSR', of class \emph{FLSR}

\dontrun{
load(neamac)

spr0=spr0Yr(neamac)
object=as.FLSR(neamac,model="bevholtSV")
sr=ftmb(object,s.est=T,s=0.7,s.logitsd=0.3,spr0)
}
}
\description{
Fits Stock Recruitment Relationships (SRR) in TBM
}
