% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\docType{methods}
\name{properties}
\alias{properties}
\title{properties}
\usage{
properties(object, ...)
}
\arguments{
\item{object}{an \code{FLBRP} object}
}
\value{
\code{FLPar} with estimates of MSY based reference points
}
\description{
Calculates additional reference points
}
\examples{
\dontrun{

data(ple4brp)

fbar(ple4brp)=FLQuant(seq(0,c(refpts(ple4brp)["crash","harvest"]),length.out=101))
refpts(ple4brp)=properties(ple4brp)

plot(ple4brp,ncol=2)
}

}
\seealso{
\code{\link{FLBRP},\link{refpts}}
}
