% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/FLBRP-pellaT.R
\name{pellaTparams}
\alias{pellaTparams}
\alias{pellaTparams,FLPar-method}
\alias{pellaTparams,FLBRP-method}
\alias{pellaTparams,numeric-method}
\title{Calculate Pella-Tomlinson Parameters}
\usage{
pellaTparams(object, ...)

\S4method{pellaTparams}{FLPar}(object, interval = c(1.1, 10))

\S4method{pellaTparams}{FLBRP}(object, interval = c(1.1, 10))

\S4method{pellaTparams}{numeric}(object, interval = c(1.1, 10))
}
\arguments{
\item{object}{containing fmsy, bmsy and b0 reference points}

\item{interval}{Numeric vector of length 2 for shape parameter optimisation bounds}
}
\value{
FLPar object with r, k, and p parameters
}
\description{
Estimates r, k, and p parameters for Pella-Tomlinson model from
MSY reference points based on an age-based model with instantaneous fishing mortality
}
