% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBRP-priors.R
\name{priors,FLBRP-method}
\alias{priors,FLBRP-method}
\title{calculates priors for biomass dynamic models}
\usage{
\S4method{priors}{FLBRP}(object)
}
\arguments{
\item{x}{An object of class FLBRP}

\item{i}{The name of the attribute to be assigned (as a character string)}

\item{value}{An object of class FLPar to be assigned as an attribute}
}
\value{
The modified FLBRP object with the new attribute
}
\description{
Assign an FLPar object as an attribute to an FLBRP object
using the '<-' operator.
}
\examples{
\dontrun{
  eql=FLBRP(...)  # Create or load an FLBRP object
  eql["priors"]<-FLPar()
  attributes(eql)$priors=calcPriors(eql)
}
setMethod("[<-", signature(object="FLBRP", i="priors", j="missing", value="FLPar"),
  function(object, i, value) {
    if (!inherits(value, "FLPar")) {
      stop("The assigned value must be an FLPar object")
    }
    
    attributes(object,i)=value

return(object)})
}
