% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{pt}
\alias{pt}
\alias{pt-method}
\alias{pt,FLPar-method}
\alias{pt,numeric-method}
\alias{pt,FLBRP-method}
\title{Production Type Method for Surplus Production Models}
\usage{
pt(object, ...)
}
\arguments{
\item{object}{Input object (FLPar, numeric, or FLBRP)}

\item{...}{Additional arguments passed to methods}

\item{shape}{Ratio of BMSY to virgin biomass (default: object\link{"bmsy"}/object\link{"v"})}

\item{what}{Function to calculate biomass metric (default: ssb)}
}
\value{
An FLPar object containing:
\itemize{
\item r: intrinsic growth rate
\item k: carrying capacity
\item p: shape parameter
\item msy: maximum sustainable yield
\item bmsy: biomass at MSY
\item fmsy: fishing mortality at MSY
\item virgin: unfished biomass
}
}
\description{
Estimates parameters of surplus production models using the PT method by deriving
parameters from reference points and biomass ratios
}
\details{
Production Type (PT) Method
}
\examples{
\dontrun{
params <- FLPar(c(bmsy=500, msy=125, v=1000))
pt(params)
}

}
