% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{rebuildTime}
\alias{rebuildTime}
\title{Calculate rebuilding time based on fish stock data.}
\usage{
rebuildTime(object, ...)
}
\arguments{
\item{stk}{An object of class \code{FLStock} representing the fish stock data.}
}
\value{
A data frame with columns "Initial" and "Year" representing the initial depletion
and the corresponding year, respectively.
}
\description{
Calculate rebuilding time based on fish stock data.
}
