% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRefs.R, R/adjust.R, R/annualise.R,
%   R/aopt.R, R/eqsim1.R, R/eqsim2.R, R/icesRefpts.R, R/nonStationarity.R,
%   R/seasonalise.R, R/spr0Cohort.R, R/spr0Yr.R
\docType{methods}
\name{addRefs}
\alias{addRefs}
\alias{adjust}
\alias{annualise}
\alias{aopt}
\alias{eqsim1}
\alias{getBlim}
\alias{icesRefpts}
\alias{nonStationarity}
\alias{seasonalise}
\alias{spr0Cohort}
\alias{spr0Yr}
\title{Adds ICES PA & MSY reference points to refpts and fits a SRR}
\usage{
addRefs(x, refs)

adjust(object)

annualise(x)

aopt(object, nyears = 3)

eqsim1(
  object,
  Blim = unlist(c(object@benchmark["Blim"])),
  Bpa = unlist(c(object@benchmark["Bpa"])),
  Btrigger = unlist(c(object@benchmark["Btrigger"])),
  Fmsy = unlist(c(object@benchmark["Fmsy"])),
  Fscan = "missing",
  Fcv = 0.212,
  Fphi = 0.423,
  bio.years = 3,
  bio.const = FALSE,
  sel.years = 3,
  sel.const = FALSE,
  verbose = FALSE,
  process.error = TRUE,
  Nrun = 200,
  nsamp = 1000,
  rhologRec = TRUE,
  SSBcv = 0,
  extreme.trim = c(0.05, 0.95)
)

getBlim(object, model = "Segreg", nsamp = 1000)

icesRefpts(x, refs = NULL, model = "bevholtSV", steepness = 0.7, nyears = 3)

nonStationarity(
  object,
  sr,
  slots = c("m", "mat", "stock.wt", "catch.wt", "catch.sel"),
  abi = FALSE
)

seasonalise(object, season = 1:4)

spr0Cohort(object)

spr0Yr(object)
}
\arguments{
\item{object}{an \code{FLStock} object}

\item{seasons}{a numeric with seasons}
}
\value{
\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object

\code{FLStock} object
}
\description{
Adds ICES PA & MSY reference points to an \code{FLStock}

Takes an \code{FLStock} that was annual and adjusts numbers-at-age to
take accoubt of seasons

Takes an \code{FLStock} with seasons and collapses into years

Age at which a cohort achieves it;s maximum biomass, i.e. where
losses due to M are greater than gains die to increase in size/mass

Runs ICES rountine to calculate MSY and PA reference points

Runs ICES rountine to calculate MSY and PA reference points

Fits an SRR to an \code{FLStock} and estimates reference points

Sets M, wt, sel & mat to vary by iter based on annual values to look at non-stationarity

While 'expand' adds seasons to an annually structured 'FLStock' it does not
change the 'FLQuants' such as m, stock.wt, mat, stock.n, catch.n and harvest
to take account of seasonal effects, such as growth and spawning. It also
does not account for changes in the population due to seasonal fishing.
Therefore 'seasonalise',  divides M into seasons, interpolates wts, and
projects a cohort across ages to estimate numbers and catch-at-age.

Calculates the spawner per recruit down a cohort

Calculates the spawner per recruit in each year
}
\examples{
\dontrun{
}
\dontrun{
}
\dontrun{
}
\dontrun{
data(ple4brp)
aopt(ple4brp)
}

\dontrun{
}
\dontrun{
}
\dontrun{
}
\dontrun{
load(neamac)

spr0=spr0Yr(neamac)

sr=as.FLSR(neamac,model="bevholtSV")
sr=ftmb(sr,s.est=T,s=0.7,s.logitsd=0.3,spr0)

refs=nonStationarity(neamac,sr)

refs=subset(refs,data>0&!is.na(data))
dat=subset(refs,(refpt=="spr.100"&quant\%in\%c("ssb"))|
                (refpt=="msy"&quant\%in\%c("ssb","harvest","rec","yield"))|
                (refpt=="virgin"&quant\%in\%c("ssb")))

ggplot(dat)+
  geom_line(aes(year,data,col=refpt))+facet_grid(quant~.,scale="free")
}
\dontrun{
data(ple4)
ple4seasonal=seasonalise(ple4)
plot(FLStocks("Annual"=ple4,"Seasonal"=ple4seasonal)
}
\dontrun{
}
\dontrun{
}
}
\seealso{
\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}

\code{\link{expand}}
}
