% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock-indicatorsBiol.R
\name{ssb.age,FLStock-method}
\alias{ssb.age,FLStock-method}
\title{ssb.age}
\usage{
\S4method{ssb.age}{FLStock}(object, ...)
}
\arguments{
\item{object}{An FLStock object containing stock assessment data.}

\item{...}{additional arguments, to replace slots in the FLStock object.}
}
\value{
Returns the spawning stock biomass by age as an FLQuant.
}
\description{
Calculates the spawning stock biomass by age for an FLStock object.
}
\examples{
data(ple4)
ssb.age(ple4)
}
\seealso{
\code{\link{ssb}}
}
