% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsBiol.R
\name{ssb,FLStock-method}
\alias{ssb,FLStock-method}
\title{ssb}
\usage{
\S4method{ssb}{FLStock}(object, ...)
}
\arguments{
\item{object}{An FLStock object containing stock assessment data.}

\item{...}{additional arguments, to replace slots in the FLStock object.}
}
\value{
Returns the spawning stock biomass as an FLQuant.
}
\description{
Calculates the spawning stock biomass for an FLStock object,
}
\examples{
data(ple4)
ssb(ple4)
}
\seealso{
\code{\link{ssb.age}}
}
