% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{invALK}
\alias{invALK}
\title{ssb}
\usage{
invALK(
  object,
  model = vonbert,
  age,
  cv = 0.1,
  lmax = 1.2,
  bin = 1,
  max = ceiling(object["linf"] * lmax),
  reflen = NULL
)
}
\arguments{
\item{object}{An FLStock object containing stock assessment data.}

\item{...}{additional arguments, to replace slots in the FLStock object.}
}
\value{
Returns the spawning stock biomass as an FLQuant.
}
\description{
Calculates the spawning stock biomass for an FLStock object,
}
\examples{
data(ple4)
ssb(ple4)
}
\seealso{
\code{\link{ssb.age}}
}
