% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{stdz}
\alias{stdz}
\alias{stdz,numeric-method}
\alias{stdz,matrix-method}
\alias{stdz,array-method}
\alias{stdz,FLQuant-method}
\title{Standardize Values}
\usage{
stdz(x, na.rm = TRUE)

\S4method{stdz}{numeric}(x, na.rm = TRUE)

\S4method{stdz}{matrix}(x, na.rm = TRUE)

\S4method{stdz}{array}(x, na.rm = TRUE)

\S4method{stdz}{FLQuant}(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector, matrix, array or FLQuant}

\item{na.rm}{Logical indicating whether to remove NA values when computing statistics}
}
\value{
An object of the same class as the input with standardized values
}
\description{
Standardizes values by subtracting the mean and dividing by the standard deviation.
}
\details{
Standardization follows the formula: (x - mean(x))/sd(x)
}
\examples{
\dontrun{
# For numeric vector
x <- 1:10
stdz(x)

# For FLQuant
data(ple4)
standardized_catch <- stdz(catch(ple4))
}

}
