% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesUtil.R
\name{tseries}
\alias{tseries}
\alias{tseries,FLStock-method}
\alias{tseries,FLStocks-method}
\title{Extract Time Series Statistics from FLStock Objects}
\usage{
tseries(object, ...)

\S4method{tseries}{FLStock}(
  object,
  flqs = list(catch = function(object) catch(object), eb = function(object)
    ebiomass(object), ssb = function(object) ssb(object), f = function(object)
    fbar(object), h = function(object) catch(object)/ebiomass(object), m =
    function(object)
    FLQuant(aaply(m(object)[ac(range(object)["minfbar"]:range(object)["maxfbar"])], 2,
    mean), dimnames = dimnames(fbar(object))))
)

\S4method{tseries}{FLStocks}(
  object,
  flqs = list(catch = function(object) catch(object), eb = function(object)
    ebiomass(object), ssb = function(object) ssb(object), f = function(object)
    fbar(object), h = function(object) catch(object)/ebiomass(object), m =
    function(object)
    FLQuant(aaply(m(object)[ac(range(object)["minfbar"]:range(object)["maxfbar"])], 2,
    mean), dimnames = dimnames(fbar(object))))
)
}
\arguments{
\item{object}{An object of class FLStock or FLStocks}

\item{...}{Additional arguments (not currently used)}
}
\value{
A data frame containing time series of:
\itemize{
\item catch - Catch values
\item eb - Exploitable biomass
\item ssb - Spawning stock biomass
\item f - Fishing mortality (Fbar)
\item h - Harvest rate (catch/ebiomass)
\item m - Mean natural mortality
}
}
\description{
Creates a data frame of time series statistics from FLStock objects, including
catch, ebiomass, SSB, fishing mortality, harvest rate, and mean natural mortality.
}
\examples{
\dontrun{
# For single stock
data(ple4)
ts1 <- tseries(ple4)

# For multiple stocks
stocks <- FLStocks(stock1=ple4, stock2=ple4)
ts2 <- tseries(stocks)
}

}
